use size_t instead of MPI_Offset for internal malloc subroutines
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/branches/am@3052 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/drivers/ncmpi/mpinetcdf.c
@@ -1156,7 +1156,7 @@ ncmpii_inq_misc(void       *ncdp,
 int ncmpi_inq_malloc_size(MPI_Offset *size)
 {
 #ifdef PNC_MALLOC_TRACE
-    ncmpii_inq_malloc_size(size);
+    ncmpii_inq_malloc_size((size_t)size);
     return NC_NOERR;
 #else
     DEBUG_RETURN_ERROR(NC_ENOTENABLED)
@@ -1169,7 +1169,7 @@ int ncmpi_inq_malloc_size(MPI_Offset *size)
 int ncmpi_inq_malloc_max_size(MPI_Offset *size)
 {
 #ifdef PNC_MALLOC_TRACE
-    ncmpii_inq_malloc_max_size(size);
+    ncmpii_inq_malloc_max_size((size_t)size);
     return NC_NOERR;
 #else
     DEBUG_RETURN_ERROR(NC_ENOTENABLED)
